The Love Letter Mystery
James got hold of a love letter that his friend Harry has written for
his girlfriend. Being the prankster that James is, he decides to meddle
with it. He changes all the words in the letter into palindromes.

While modifying the letters of the word, he follows 2 rules:

(a) He always reduces the value of a letter, e.g. he changes 'd' to 'c',
but he does not change 'c' to 'd'.
(b) If he has to repeatedly reduce the value of a letter, he can do it
until the letter becomes 'a'. Once a letter has been changed to 'a', it
can no longer be changed.

Each reduction in the value of any letter is counted as a single
operation. Find the minimum number of operations he carries out to
convert a given string into a palindrome.